requirements.txt
django
django-rest-framework
pygithub
django-redis
